expanded description for "radius"
I originally understood the wording around "65535 = pill shaped widget (true circle..." in a way that setting radius to 65535 always gives you a circle, also for widgets that (obviously) have way smaller widths and heights. Working with it a bit I realised it's not that straightforward, and that setting a radius in line with what is set for width/height is the way to go anyway. So I decided to expand the description a bit.